+++ R/flexscan.R
@@ -154,6 +154,9 @@ flexscan.rantype <- c("MULTINOMIAL", "POISSON")
 #'   \item{"BOTH"}{Hot- and cold-spot clusters simultaneously.}
 #' }
 #' 
+#' @param clusterradius
+#' The maximum radius of spatial cluster to scan.
+#' 
 #' @return 
 #' An \code{rflexscan} object which contains analysis results and specified
 #' parameters.
@@ -222,7 +225,8 @@ rflexscan <- function(x, y, lat, lon,
                       comments="",
                       verbose=FALSE,
                       secondary=NULL,
-                      clustertype="HOT") {
+                      clustertype="HOT",
+                      clusterradius=.Machine$double.xmax) {
   call <- match.call()
 
   stattype <- match.arg(toupper(stattype), flexscan.stattype)
@@ -288,14 +292,16 @@ rflexscan <- function(x, y, lat, lon,
   setting$secondary <- ifelse(is.null(secondary), -1, secondary)
 
   if (toupper(clustertype) == "HOT") {
-    setting$clustertype = 1
+    setting$clustertype <- 1
   } else if (toupper(clustertype) == "COLD") {
-    setting$clustertype = 2
+    setting$clustertype <- 2
   } else if (toupper(clustertype) == "BOTH") {
-    setting$clustertype = 3
+    setting$clustertype <- 3
   } else {
-    setting$clustertype = 1
+    setting$clustertype <- 1
   }
+  
+  setting$clusterradius <- clusterradius
   
   if (!verbose) {
     output <- capture.output({
@@ -359,6 +365,10 @@ rflexscan <- function(x, y, lat, lon,
 #' @param ...
 #' Ignored.
 #' 
+#' @return
+#' No return value, called for side effects.
+#' Prints a concise summary of the detected clusters and model settings.
+#' 
 #' @seealso \link{rflexscan}
 #' 
 #' @method print rflexscan
@@ -383,6 +393,11 @@ print.rflexscan <- function(x, ...) {
 #' 
 #' @param ...
 #' Ignored.
+#' 
+#' @return
+#' No return value, called for side effects.
+#' Prints details of a single detected cluster, including member regions and 
+#' statistics.
 #' 
 #' @method print rflexscanCluster
 #' @export
@@ -417,6 +432,11 @@ print.rflexscanCluster <- function(x, ...) {
 #' 
 #' @param ...
 #' Ignored.
+#' 
+#' @return
+#' An object of class \code{"summary.rflexscan"} including summary tables of 
+#' detected clusters, test statistics, and p-values. This object can be printed 
+#' with \code{\link{print.summary.rflexscan}}.
 #' 
 #' @seealso \link{rflexscan}
 #' 
@@ -468,6 +488,11 @@ summary.rflexscan <- function(object, ...) {
 #' @param ...
 #' Ignored.
 #' 
+#' @return
+#' No return value, called for side effects.
+#' Prints the summarized results of the flexible spatial scan statistic, 
+#' including the number of clusters and test statistics.
+#' 
 #' @seealso \link{rflexscan}, \link{summary.rflexscan}
 #' 
 #' @method print summary.rflexscan
@@ -507,6 +532,7 @@ print.summary.rflexscan <- function(x, ...) {
   cat("---\nSignif. codes: ", attr(signif, "legend"), "\n\n")
   
   cat("Limit length of cluster:", x$setting$clustersize, "\n")
+  cat("Limit radius of cluster:", x$setting$clusterradius, "\n")
   cat("Number of areas:", x$total_areas, "\n")
   cat("Total cases:", x$total_cases, "\n")
   if (x$setting$cartesian) {
@@ -559,7 +585,11 @@ print.summary.rflexscan <- function(x, ...) {
 #' Fill color of vertices that are not included in any clusters.
 #' 
 #' @param ...
-#' Other parameters to be passed to \link{plot.igraph} function.
+#' Other parameters to be passed to \link[igraph:plot.igraph]{plot.igraph} function.
+#' 
+#' @return 
+#' No return value, called for side effects.
+#' Displays the detected clusters on a graph.
 #' 
 #' @details 
 #' Clusters are colored using the current palette. Please use \link{palette}
@@ -662,6 +692,10 @@ plot.rflexscan <- function(x,
 #' @param ...
 #' Other parameters to be passed to plot function.
 #' 
+#' @return 
+#' No return value, called for side effects. 
+#' This function generates a choropleth map of cluster results.
+#' 
 #' @details 
 #' Clusters are colored using the current palette. Please use \link{palette}
 #' function to specify colors of each cluster. Note that clusters with ranks
@@ -675,7 +709,7 @@ plot.rflexscan <- function(x,
 #' library(sf)
 #' library(spdep)
 #' data("nc.sids")
-#' sids.shp <- read_sf(system.file("shapes/sids.shp", package="spData")[1])
+#' sids.shp <- st_read(system.file("shapes/sids.gpkg", package="spData")[1], quiet=TRUE)
 #' 
 #' # calculate the expected numbers of cases
 #' expected <- nc.sids$BIR74 * sum(nc.sids$SID74) / sum(nc.sids$BIR74)